+++ features/fixtures/maze_runner/Assets/Scripts/Scenarios/Csharp/Config/Context.cs
@@ -1,0 +1,13 @@
+﻿public class Context : Scenario
+{
+    public override void PrepareConfig(string apiKey, string host)
+    {
+        base.PrepareConfig(apiKey, host);
+        Configuration.Context = "test";
+    }
+
+    public override void Run()
+    {
+        DoSimpleNotify("Context");
+    }
+}