+++ features/fixtures/maze_runner/Assets/Scripts/Scenarios/Csharp/Config/LongLaunchTime.cs
@@ -1,0 +1,26 @@
+﻿using BugsnagUnity;
+
+public class LongLaunchTime : Scenario
+{
+    public override void PrepareConfig(string apiKey, string host)
+    {
+        base.PrepareConfig(apiKey, host);
+        Configuration.LaunchDurationMillis = 7000;
+    }
+
+    public override void Run()
+    {
+        Invoke("DoNotify1",6);
+        Invoke("DoNotify2", 8);
+    }
+
+    private void DoNotify1()
+    {
+        Bugsnag.Notify(new System.Exception("Error 1"));
+    }
+
+    private void DoNotify2()
+    {
+        Bugsnag.Notify(new System.Exception("Error 2"));
+    }
+}