+++ features/fixtures/maze_runner/Assets/Scripts/Scenarios/Csharp/Config/ReleaseStage.cs
@@ -1,0 +1,16 @@
+﻿using BugsnagUnity;
+
+public class ReleaseStage : Scenario
+{
+    public override void PrepareConfig(string apiKey, string host)
+    {
+        base.PrepareConfig(apiKey, host);
+        Configuration.EnabledReleaseStages = new string[] {"release"};
+        Configuration.ReleaseStage = "notRelease";
+    }
+
+    public override void Run()
+    {
+        Bugsnag.Notify(new System.Exception("ReleaseStage"));
+    }
+}