+++ features/fixtures/maze_runner/Assets/Scripts/Scenarios/Csharp/Events/CustomSeverity.cs
@@ -1,0 +1,9 @@
+﻿using BugsnagUnity;
+
+public class CustomSeverity : Scenario
+{
+    public override void Run()
+    {
+        Bugsnag.Notify(new System.Exception("CustomSeverity"), Severity.Info);
+    }
+}