+++ features/fixtures/maze_runner/Assets/Scripts/Scenarios/Csharp/Events/DebugLogWarningSmokeTest.cs
@@ -1,0 +1,16 @@
+﻿using UnityEngine;
+
+public class DebugLogWarningSmokeTest : Scenario
+{
+    public override void PrepareConfig(string apiKey, string host)
+    {
+        base.PrepareConfig(apiKey, host);
+        Configuration.NotifyLogLevel = LogType.Warning;
+    }
+
+    public override void Run()
+    {
+        AddTestingMetadata();
+        Debug.LogWarning("DebugLogWarningSmokeTest");
+    }
+}