+++ features/fixtures/maze_runner/Assets/Scripts/Scenarios/Csharp/Events/NotifyWithCustomStacktrace.cs
@@ -1,0 +1,10 @@
+﻿using BugsnagUnity;
+using System;
+
+public class NotifyWithCustomStacktrace : Scenario
+{
+    public override void Run()
+    {
+        Bugsnag.Notify(new Exception("NotifyWithCustomStacktrace"), CustomStacktrace);        
+    }
+}