+++ features/fixtures/maze_runner/Assets/Scripts/Scenarios/Csharp/Sessions/NewSession.cs
@@ -1,0 +1,21 @@
+﻿using System.Collections;
+using System.Collections.Generic;
+using BugsnagUnity;
+using UnityEngine;
+
+public class NewSession : Scenario
+{
+    public override void PrepareConfig(string apiKey, string host)
+    {
+        base.PrepareConfig(apiKey, host);
+        Configuration.AutoTrackSessions = false;
+    }
+
+    public override void Run()
+    {
+        Bugsnag.StartSession();
+        Bugsnag.Notify(new System.Exception("Error 1"));
+        Bugsnag.StartSession();
+        Bugsnag.Notify(new System.Exception("Error 2"));
+    }
+}