+++ features/fixtures/maze_runner/Assets/Scripts/Scenarios/Csharp/User/SetUserAfterStart.cs
@@ -1,0 +1,18 @@
+﻿using BugsnagUnity;
+using System;
+
+public class SetUserAfterStart : Scenario
+{
+
+    public override void PrepareConfig(string apiKey, string host)
+    {
+        base.PrepareConfig(apiKey, host);
+    }
+
+    public override void Run()
+    {
+        Bugsnag.SetUser("1", "2", "3");
+        Bugsnag.Notify(new Exception("SetUserAfterStart"));
+    }
+
+}