+++ features/fixtures/maze_runner/Assets/Scripts/Scenarios/MacOS/MacOSNativeCrashCallback.cs
@@ -1,0 +1,15 @@
+﻿using System.Collections;
+using System.Collections.Generic;
+using UnityEngine;
+
+public class MacOSNativeCrashCallback : Scenario
+{
+
+    public override void PrepareConfig(string apiKey, string host)
+    {
+        base.PrepareConfig(apiKey, host);
+        Configuration.AddOnSendError(CocoaNativeEventCallback);
+    }
+
+  
+}