+++ features/fixtures/maze_runner/Assets/Scripts/Scenarios/MacOS/MacOSNativeCrashOutsideReleaseStages.cs
@@ -1,0 +1,18 @@
+﻿using System.Collections;
+using System.Collections.Generic;
+using UnityEngine;
+
+public class MacOSNativeCrashOutsideReleaseStages : Scenario
+{
+    public override void PrepareConfig(string apiKey, string host)
+    {
+        base.PrepareConfig(apiKey, host);
+        Configuration.EnabledReleaseStages = new string[] { "production" };
+        Configuration.ReleaseStage = "notProduction";
+    }
+
+    public override void Run()
+    {
+        MacOSNativeCrash();
+    }
+}